factor away the base storage class
